requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langgraph==0.0.26
google-generativeai==0.3.2
tavily-python==0.3.0
pydantic==2.5.0
python-multipart==0.0.6